+++ src/main/java/com/swifty/bank/server/api/controller/dto/auth/response/CreateSecureKeypadResponse.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.api.controller.dto.keypad.response;
+package com.swifty.bank.server.api.controller.dto.auth.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
